backend/requirements-minimal.txt
# Core requirements
numpy==1.26.4
pandas==2.2.2
sqlalchemy==2.0.20
nltk==3.8.1
scipy==1.10.1

# Essential ML requirements for Docker
torch==2.2.2
transformers==4.36.2
scikit-learn==1.4.0

# Core NLP requirements (minimal)
spacy==3.6.1

# Text extraction requirements
PyPDF2==3.0.1
python-docx==1.1.0
tika==2.6.0

# Web and API requirements
fastapi==0.109.0
uvicorn[standard]==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
python-magic==0.4.27

# Background processing
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Testing requirements
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
reportlab==4.0.7

# Development requirements
black==23.7.0
isort==5.12.0
flake8==6.1.0